+++ components/Form/Checkbox.jsx
@@ -15,6 +15,7 @@ const CheckboxGroup = ({
   direction = 'horizontal',
   style,
   className,
+  virtual,
   ...props
 }) => {
 
@@ -35,9 +36,13 @@ const CheckboxGroup = ({
   }, [onChange, value, disabled])
 
   return <context.Provider value={{ check, currentValue: value }}>
-    <Space row={horizontal} items={horizontal ? 'center' : 'stretch'} wrap={horizontal} {...props} style={style} className={classNames('flex-grow', className)}>
-      {children}
-    </Space>
+    {
+      virtual ?
+        children :
+        <Space row={horizontal} items={horizontal ? 'center' : 'stretch'} wrap={horizontal} {...props} style={style} className={classNames('flex-grow', className)}>
+          {children}
+        </Space>
+    }
   </context.Provider>
 }