+++ components/HtmlView/components/RichText/index.js
@@ -1,4 +1,5 @@
 import { RichText as TaroRichText } from '@tarojs/components'
+import './index.scss'
 
 export const RichText = ({ style, ...props }) => {
   return <TaroRichText style={{ fontSize: '16px', ...style }} {...props} />